Adding Mixture and Tablet classes
