+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/database/data/Antibiotic.java
@@ -2,35 +2,27 @@ package fi.tuni.koodimankelit.antibiootit.database.data;
 
 import java.util.List;
 
-public class Antibiotic {
+public abstract class Antibiotic {
     private final String antibiotic;
     private final String format;
     private final String info;
-    private final String dosagePerDayUnit;
-    private final int dosagePerDay;
     private final int maxDosePerDay;
     private final List<Strength> strength;
     private final String weightUnit;
-    private final String resultUnit;
     private final int days;
     private final int dosesPerDay;
-    private final int tabletsPerDose;
 
-    public Antibiotic(String antibiotic, String format, String info, String dosagePerDayUnit, int dosagePerDay, 
-        int maxDosePerDay, List<Strength> strength, String weightUnit, String resultUnit, int days,
-        int dosesPerDay, int tabletsPerDose) {
+    public Antibiotic(String antibiotic, String format, String info, 
+        int maxDosePerDay, List<Strength> strength, String weightUnit, int days,
+        int dosesPerDay) {
             this.antibiotic = antibiotic;
             this.format = format;
             this.info = info;
-            this.dosagePerDayUnit = dosagePerDayUnit;
-            this.dosagePerDay = dosagePerDay;
             this.maxDosePerDay = maxDosePerDay;
             this.strength = strength;
             this.days = days;
             this.dosesPerDay = dosesPerDay;
             this.weightUnit = weightUnit;
-            this.resultUnit = resultUnit;
-            this.tabletsPerDose = tabletsPerDose;
     }
 
     public String getAntibiotic() {
@@ -45,14 +37,6 @@ public class Antibiotic {
         return this.info;
     }
 
-    public String getDosagePerDayUnit() {
-        return this.dosagePerDayUnit;
-    }
-
-    public int getDosagePerDay() {
-        return this.dosagePerDay;
-    }
-
     public int getMaxDosePerDay() {
         return this.maxDosePerDay;
     }
@@ -65,19 +49,11 @@ public class Antibiotic {
         return this.weightUnit;
     }
 
-    public String getResultUnit() {
-        return this.resultUnit;
-    }
-
     public int getDays() {
         return this.days;
     }
 
     public int getDosesPerDay() {
         return this.dosesPerDay;
-    }
-
-    public int getTabletsPerDose() {
-        return this.tabletsPerDose;
     }
 }
